Add type definitions for punycode.js
Since January 19th ([1]), the `punycode` package has also been published
under `punycode.js`. This solves the problem of needing to use
`punycode/` to use the package (as `punycode` is a built-in Node
package). However, types are missing for `punycode.js`. So this commit
adds them. Note that the interface for the package hasn't changed since
v2.1, so I didn't bump the version, just copied the existing type
definitions from `punycode`.

[1]: https://github.com/mathiasbynens/punycode.js/issues/122